Robustness fixes.
Needed because client-side testcards still issue http psychic requests and
that's difficult to fix on the client-side. So, just sanely returning nothing
is the best option.

Pretty soon there should be no client-side testcards so this problem will
go away, but to migrate them to the server-side does require them working
correctly.

And robustness is never a bad thing.